backend/cars/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
mysql==0.0.2
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
environ==1.0